requirements.txt
opencv-python
numpy
Pillow
pygame

# Raspberry Pi only (not needed for desktop simulation):
# luma.oled
# luma.core
# gpiozero
